Fetch bitmap from inputstream in DataFetcher#loadData
