+++ app/src/main/java/com/amaze/filemanager/utils/glide/CloudIconModelLoader.java
@@ -1,0 +1,41 @@
+package com.amaze.filemanager.utils.glide;
+
+import android.content.Context;
+import android.support.annotation.Nullable;
+
+import com.amaze.filemanager.database.CloudHandler;
+import com.bumptech.glide.load.Options;
+import com.bumptech.glide.load.model.ModelLoader;
+import com.bumptech.glide.signature.ObjectKey;
+
+/**
+ * Created by Vishal Nehra on 3/27/2018.
+ */
+
+public class CloudIconModelLoader implements ModelLoader<String, String> {
+
+    private Context context;
+
+    public CloudIconModelLoader(Context context) {
+        this.context = context;
+    }
+
+    @Nullable
+    @Override
+    public LoadData<String> buildLoadData(String s, int width, int height, Options options) {
+        // we put key as current time since we're not disk caching the images for cloud,
+        // as there is no way to differentiate input streams returned by different cloud services
+        // for future instances and they don't expose concrete paths either
+        return new LoadData(new ObjectKey(System.currentTimeMillis()), new CloudIconDataFetcher(context, s));
+    }
+
+    @Override
+    public boolean handles(String s) {
+        return s.startsWith(CloudHandler.CLOUD_PREFIX_BOX)
+                || s.startsWith(CloudHandler.CLOUD_PREFIX_DROPBOX)
+                || s.startsWith(CloudHandler.CLOUD_PREFIX_GOOGLE_DRIVE)
+                || s.startsWith(CloudHandler.CLOUD_PREFIX_ONE_DRIVE)
+                || s.startsWith("smb:/")
+                || s.startsWith("sftp:/");
+    }
+}
